fix(meetings): address CodeRabbit review on PR #298
Fix a false-conflict bug: a count-bounded recurrence (numberOfOccurrences
set, no explicit endDate) was checked against its raw null endDate,
expanding it out to the full 2-year overlap horizon and risking a false
409 against a booking that falls after the series' real end.
calculateEndDateFromOccurrences is now shared (util/meetingOccurrences.ts)
so both routes resolve the same finite endpoint before checking.

Fix a related suspension bug in findResourceConflictRows: filtering by
whether a meeting is suspended today (not per-occurrence) either missed
a real conflict (meeting resumes before a future candidate occurrence)
or produced a false one (meeting has a future suspension window). Now
filtered per-occurrence date, matching computeConflicts' own semantics.

Give ConflictOverrideModal real dialog semantics (role, aria-modal,
initial focus, Escape-to-cancel, focus restoration) mirroring
BottomSheet's existing pattern -- previously focus stayed on the
form's own submit button underneath, risking a double-submit. Add
max-height/scroll so the modal can't get clipped on a short viewport.

Break a circular import (ConflictList -> EditMeeting -> ConflictOverride
Modal -> ConflictList) by extracting the shared conflict types/formatting
helpers into util/conflictDisplay.ts; ConflictList re-exports them for
existing callers.

Switch two fixed-delay integration tests to poll for an observable
completion state with a real assertion, and add coverage for the
suspension fix's two directions (resumes-before-candidate,
future-suspension-window).
