+++ frontend/app/api/write/meeting/route.ts
@@ -5,8 +5,8 @@ import { requireRole } from "../../../../services/auth";
 import { createCalendarEvent, calendarIdsForMeeting } from "../../../../services/googleCalendar";
 import { createZoomMeeting, getZoomMeetingInvitation, resolveZoomHost, zoomRoomCalendarId } from "../../../../services/zoom";
 import { findResourceConflicts, findResourceConflictRows, ConflictRow } from "../../../../util/resourceOverlap";
-import { convertETToUTC } from "../../../../util/timeUtils";
 import { meetingSchema } from "../../../../util/meetingValidation";
+import { calculateEndDateFromOccurrences } from "../../../../util/meetingOccurrences";
 import { prisma } from "../../../../lib/prisma";
 
 // Runs after the response is sent (see after() call below) — failure sets googleSyncStatus but
@@ -146,11 +146,33 @@ const createMeeting = async (request: Request) => {
     const { recurrencePattern, confirmOverride, ...meetingDetails } = meetingData;
     const isRecurring = !!recurrencePattern;
 
+    // Resolved once, up front, so the conflict check below and the eventual RecurrencePattern
+    // write use the same finite endpoint -- a count-bounded series (numberOfOccurrences set, no
+    // explicit endDate) has a real last occurrence, and checking conflicts against the raw
+    // (still-null) endDate would expand it out to the full OVERLAP_HORIZON_YEARS window instead,
+    // risking a false 409 against an unrelated booking that falls after the series actually ends.
+    let calculatedEndDate = recurrencePattern?.endDate ?? null;
+    if (recurrencePattern?.numberOfOccurrences && !recurrencePattern.endDate) {
+      calculatedEndDate = calculateEndDateFromOccurrences(
+        recurrencePattern.startDate,
+        recurrencePattern.daysOfWeek || [],
+        recurrencePattern.numberOfOccurrences,
+        recurrencePattern.interval || 1,
+        recurrencePattern.type,
+        recurrencePattern.weekOfMonth ?? null,
+        recurrencePattern.dayOfMonth ?? null,
+      );
+    }
+
     // Blocks the save outright on a room/zoomRoom collision -- distinct from the zoomHost check
     // below, which defers the calendar publish and stores the error on the meeting instead of
     // rejecting the request. confirmOverride only bypasses this block, not zoomHost's handling.
     if (!confirmOverride) {
-      const candidate = { ...meetingData, isRecurring };
+      const candidate = {
+        ...meetingData,
+        isRecurring,
+        recurrencePattern: recurrencePattern ? { ...recurrencePattern, endDate: calculatedEndDate } : null,
+      };
       const conflictRows: ConflictRow[] = [];
       if (meetingData.room) {
         conflictRows.push(...await findResourceConflictRows("room", meetingData.room, candidate, { excludeMid: meetingData.mid }));
@@ -206,20 +228,6 @@ const createMeeting = async (request: Request) => {
     let responseMeeting: object;
 
     if (recurrencePattern) {
-      let calculatedEndDate = recurrencePattern.endDate;
-
-      if (recurrencePattern.numberOfOccurrences && !recurrencePattern.endDate) {
-        calculatedEndDate = calculateEndDateFromOccurrences(
-          recurrencePattern.startDate,
-          recurrencePattern.daysOfWeek || [],
-          recurrencePattern.numberOfOccurrences,
-          recurrencePattern.interval || 1,
-          recurrencePattern.type,
-          recurrencePattern.weekOfMonth ?? null,
-          recurrencePattern.dayOfMonth ?? null,
-        );
-      }
-
       // meetingDetails.mid is client-generated (see NewMeeting.tsx's uuidv4()) and known before
       // either write, so both creates can run as one atomic transaction instead of a create-then-
       // create sequence an interrupted request could leave half-done (a Meeting with
@@ -261,109 +269,5 @@ const createMeeting = async (request: Request) => {
     return NextResponse.json({ error: "Internal Server Error" }, { status: 500 });
   }
 };
-
-/**
- * Calculate the end date based on a specific number of occurrences
- */
-function calculateEndDateFromOccurrences(
-  startDate: Date,
-  daysOfWeek: string[],
-  numberOfOccurrences: number,
-  interval: number,
-  type: string,
-  weekOfMonth: number | null = null,
-  dayOfMonth: number | null = null,
-): Date {
-  const patternStartDate = new Date(startDate);
-
-  if (numberOfOccurrences <= 0) return patternStartDate;
-
-  const dayNameToIndex: Record<string, number> = {
-    "Sunday": 0, "Monday": 1, "Tuesday": 2, "Wednesday": 3,
-    "Thursday": 4, "Friday": 5, "Saturday": 6,
-  };
-
-  if (type === "monthly") {
-    // The Nth occurrence is (N-1) intervals after the start month
-    const rawMonth = patternStartDate.getUTCMonth() + (numberOfOccurrences - 1) * interval;
-    const targetYear = patternStartDate.getUTCFullYear() + Math.floor(rawMonth / 12);
-    const targetMonth = rawMonth % 12;
-
-    // 23:59:59 ET so the end date is inclusive of its full day
-    // even against a naive instant comparison (e.g. `meetingStart <= endDate`).
-    const toETDate = (day: number) => new Date(convertETToUTC(
-      `${targetYear}-${String(targetMonth + 1).padStart(2, '0')}-${String(day).padStart(2, '0')}T23:59:59`
-    ));
-
-    if (dayOfMonth != null) {
-      return toETDate(dayOfMonth);
-    }
-
-    if (weekOfMonth != null && daysOfWeek.length > 0) {
-      const targetDay = dayNameToIndex[daysOfWeek[0]];
-      const daysInMonth = new Date(Date.UTC(targetYear, targetMonth + 1, 0)).getUTCDate();
-
-      if (weekOfMonth === -1) {
-        for (let d = daysInMonth; d >= 1; d--) {
-          if (new Date(Date.UTC(targetYear, targetMonth, d)).getUTCDay() === targetDay) {
-            return toETDate(d);
-          }
-        }
-      } else {
-        let count = 0;
-        for (let d = 1; d <= daysInMonth; d++) {
-          if (new Date(Date.UTC(targetYear, targetMonth, d)).getUTCDay() === targetDay) {
-            if (++count === weekOfMonth) {
-              return toETDate(d);
-            }
-          }
-        }
-      }
-    }
-
-    return toETDate(patternStartDate.getUTCDate());
-  }
-
-  // Weekly
-  if (daysOfWeek.length === 0) return patternStartDate;
-
-  const recurrenceDays = daysOfWeek
-    .map(day => dayNameToIndex[day])
-    .filter(index => index !== undefined)
-    .sort((a, b) => a - b);
-
-  if (recurrenceDays.length === 0) return patternStartDate;
-
-  const endDate = new Date(patternStartDate);
-  let occurrenceCount = 0;
-  let currentWeek = 0;
-  const startDayOfWeek = patternStartDate.getUTCDay();
-
-  // The start date only counts as an occurrence if its weekday is in daysOfWeek.
-  if (recurrenceDays.includes(startDayOfWeek)) {
-    occurrenceCount++;
-    if (occurrenceCount >= numberOfOccurrences) return patternStartDate;
-  }
-
-  let nextDayIndex = recurrenceDays.findIndex(day => day > startDayOfWeek);
-  if (nextDayIndex === -1) { nextDayIndex = 0; currentWeek++; }
-
-  while (occurrenceCount < numberOfOccurrences) {
-    if (currentWeek % interval === 0) {
-      while (nextDayIndex < recurrenceDays.length) {
-        const daysToAdd = (currentWeek * 7) +
-          (recurrenceDays[nextDayIndex] - startDayOfWeek + 7) % 7;
-        endDate.setUTCDate(patternStartDate.getUTCDate() + daysToAdd);
-        occurrenceCount++;
-        nextDayIndex++;
-        if (occurrenceCount >= numberOfOccurrences) return endDate;
-      }
-    }
-    currentWeek++;
-    nextDayIndex = 0;
-  }
-
-  return endDate;
-}
 
 export { createMeeting as POST };
