+++ frontend/app/components/meeting-form/EditMeeting.tsx
@@ -16,7 +16,7 @@ import IconButton from '@mui/material/IconButton';
 import { IMeeting } from '../../../util/models'
 import { physicalRoomOptions, zoomRoomOptions } from "../../../util/rooms";
 import { useMeetingForm, CAL_TYPE_OPTIONS, CAL_TYPE_COLOR, DESCRIPTION_MAX_LENGTH } from '../../../hooks/useMeetingForm';
-import { ConflictListRow } from '../admin/ConflictList';
+import { ConflictListRow } from '../../../util/conflictDisplay';
 
 import styles from '../../../styles/components/meeting-form/MeetingForm.module.scss';
